+++ llrt_core/src/utils/mod.rs
@@ -3,6 +3,7 @@
 pub mod clone;
 pub mod io;
 pub mod mc_oneshot;
+pub mod object;
 pub mod result {
     pub use llrt_utils::result::*;
 }
