+++ llrt_utils/src/object.rs
@@ -1,11 +1,6 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
-use std::collections::{BTreeMap, HashMap};
-
-use rquickjs::{
-    atom::PredefinedAtom, Array, Coerced, Ctx, FromJs, Function, IntoAtom, IntoJs, Object, Result,
-    Symbol, Value,
-};
+use rquickjs::{atom::PredefinedAtom, FromJs, Function, IntoAtom, Object, Result, Symbol, Value};
 
 pub trait ObjectExt<'js> {
     fn get_optional<K: IntoAtom<'js> + Clone, V: FromJs<'js>>(&self, k: K) -> Result<Option<V>>;
@@ -39,54 +34,4 @@ impl<'js> CreateSymbol<'js> for Symbol<'js> {
         let for_function: Function = symbol_function.get(PredefinedAtom::For)?;
         for_function.call((description,))
     }
-}
-
-#[allow(dead_code)]
-pub fn array_to_hash_map<'js>(
-    ctx: &Ctx<'js>,
-    array: Array<'js>,
-) -> Result<HashMap<String, String>> {
-    let value = object_from_entries(ctx, array)?;
-    let value = value.into_value();
-    HashMap::from_js(ctx, value)
-}
-
-pub fn array_to_btree_map<'js>(
-    ctx: &Ctx<'js>,
-    array: Array<'js>,
-) -> Result<BTreeMap<String, Coerced<String>>> {
-    let value = object_from_entries(ctx, array)?;
-    let value = value.into_value();
-    BTreeMap::from_js(ctx, value)
-}
-
-pub fn object_from_entries<'js>(ctx: &Ctx<'js>, array: Array<'js>) -> Result<Object<'js>> {
-    let obj = Object::new(ctx.clone())?;
-    for value in array.into_iter().flatten() {
-        if let Some(entry) = value.as_array() {
-            if let Ok(key) = entry.get::<Value>(0) {
-                if let Ok(value) = entry.get::<Value>(1) {
-                    let _ = obj.set(key, value); //ignore result of failed
-                }
-            }
-        }
-    }
-    Ok(obj)
-}
-
-pub fn map_to_entries<'js, K, V, M>(ctx: &Ctx<'js>, map: M) -> Result<Array<'js>>
-where
-    M: IntoIterator<Item = (K, V)>,
-    K: IntoJs<'js>,
-    V: IntoJs<'js>,
-{
-    let array = Array::new(ctx.clone())?;
-    for (idx, (key, value)) in map.into_iter().enumerate() {
-        let entry = Array::new(ctx.clone())?;
-        entry.set(0, key)?;
-        entry.set(1, value)?;
-        array.set(idx, entry)?;
-    }
-
-    Ok(array)
 }
